Module Fuseau.Event_loop

Abstraction over an event loop, from the pov of the scheduler.

All the scheduler cares about is running an iteration of the loop, with or without blocking; and interrupting the loop if it's in the middle of blocking.

class type t = object ... end

Abstract event loop, inspired by Lwt engine

val one_step : t -> block:bool -> unit -> unit

val on_timer : t -> float -> repeat:bool -> (Cancel_handle.t -> unit) -> Cancel_handle.t

val interrupt_if_in_blocking_section : t -> unit