Class type Event_loop.t

Abstract event loop, inspired by Lwt engine

method one_step : block:bool -> unit -> unit

Run one step of the event loop.

parameter block
if true, the call might block until the next timeout or until the next IO event occurs. If false, this does not block and returns after having processed the available events.

method on_timer : float -> repeat:bool -> (Fuseau__.Cancel_handle.t -> unit) -> Fuseau__.Cancel_handle.t

on_timer delay ~repeat f runs f after delay.

parameter repeat
if true runs f every delay seconds

method interrupt_if_in_blocking_section : unit

If run from inside the event loop when it's waiting, wakes the event loop up