Module Fuseau.Iostream

IO streams.

These streams of bytes can be used for read and write operations. Their usage is not required, in some event loops such as Fuseau_lwt.

module Out : sig ... end

Output stream.

module In : sig ... end

Input stream.