Module Fuseau.Scheduler

Scheduler that runs fibers.

The scheduler is responsible for running fibers that are ready, but it doesn't directly deal with timeouts, readiness events, etc. For these see Event_loop.

type t

A scheduler

val active : t -> bool

Is the scheduler not finished yet?

val dispose : t -> unit

Delete the scheduler. Idempotent and thread-safe. This cancels the remaining fibers. The scheduler will stop running when they all terminate.

val n_tasks_since_beginning : t -> int

Number of tasks run so far in this scheduler.